+++ lib/features/home/widgets/side_control_block.dart
@@ -1,3 +1,4 @@
+import 'package:ebalistyka/shared/helpers/debug_highlight.dart';
 import 'package:flutter/material.dart';
 
 class SideControlBlock extends StatelessWidget {