+++ lib/shared/widgets/help_dialog.dart
@@ -16,6 +16,18 @@ class HelpData {
   static const String targetDistanceConvertor = 'targetDistanceConvertor';
   static const String angularConvertor = 'angularConvertor';
   static const String reticleScreen = 'reticleScreen';
+  static const String profilesScreen = 'profilesScreen';
+  static const String weaponCollectionScreen = 'weaponCollectionScreen';
+  static const String ammoCollectionScreen = 'ammoCollectionScreen';
+  static const String sightCollectionScreen = 'sightCollectionScreen';
+  static const String mySightsScreen = 'mySightsScreen';
+  static const String myAmmoScreen = 'myAmmoScreen';
+  static const String multiBcEditor = 'multiBcEditor';
+  static const String customDragEditor = 'customDragEditor';
+  static const String powderSensEditor = 'powderSensEditor';
+  static const String sightWizard = 'sightWizard';
+  static const String weaponWizard = 'weaponWizard';
+  static const String ammoWizard = 'ammoWizard';
 
   final String title;
   final String body;